feat(package)!: Implements package instal
This PR implements installation as a package of the repo and instruction on how to install it.
Some additional changes on the pathing were required to make it compatible but I tried my best to only change what was required.
