+++ README.md
@@ -4,6 +4,9 @@ Chronologer is a deep learning model for highly accurate prediction of peptide C
 
 
 # Installation
+
+## Option 1
+
 Clone the github repository into the desired directory. 
 
 Chronologer has been verified to work under the following environment:
@@ -15,6 +18,32 @@ Chronologer has been verified to work under the following environment:
 | Pandas | 1.4.3 |
 | Scipy | 1.9.1 |
 | Pyteomics | 4.5.5 |
+
+## Option 2
+
+It is highly encouraged to use virtual environments to
+isolate the package.
+
+From the command line this can be done like so:
+
+```shell
+python3 -m venv venv # creates an environment called venv
+source venv/bin/activate # activates it
+```
+
+The package can then be installed from
+github directly like so:
+
+```
+python -m pip install "chronologer @ git+https://github.com/searlelab/chronologer.git"
+```
+
+And we can check if it installed correctly using the
+chronologer command.
+
+```
+chronologer_predict --help
+```
 
 # Usage
 Chronologer can predict C18 retention coefficients for peptides between 6-50 amino acids in length and any of the supported modifications:
@@ -49,8 +78,11 @@ The retention time prediction script (Predict_RT.py) takes as its input a tab-se
 Chronologer can be re-trained to support additional modifications (mininum 10 peptides). Re-training first requires alignment of user retention times into Hydrophobic Index (HI) space (i.e. the % ACN in 0.1% FA that a peptide will elute). As long as user source data contains some Chronologer-predictable peptides (preferably >100), this can be easily done using the provided Align_RT_to_Hydrophobic_Index.py script (use -h flag for detailed instructions). To train Chronologer for new modifications, the modification masses must first be defined in the user_modifications.txt file, which is a TSV file with 3 columns: index, residue, and mass. Do not alter the index column, which is used for internal purposes. User provided masses must perfectly match the values listed in the PeptideModSeq strings. Modifications can be disabled by commenting out the line (i.e. adding # to the beginning of the line). Detailed instructions for training Chronologer can be found by running "python Train_Chronologer.py -h". 
 
 # Demo
+
 A small test file (demo_peptides.txt) with 100 "crap peptides" is included to test a working environment. From the command line, retention times can be predicted by running
-### python Predict_RT.py demo_peptides.txt demo_rt.txt
+
+> python Predict_RT.py demo_peptides.txt demo_rt.txt
+
 where "demo_rt.txt" can be replaced with the output file name of your choice
 
 # Citation
