+++ pyproject.toml
@@ -1,0 +1,33 @@
+[build-system]
+requires = ["setuptools"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "chronologer"
+version = "0.2.0"
+authors = [
+    {name = "Damien Beau Wilburn"},
+    {name = "Brian C. Searle"}
+]
+description = ""
+readme = "README.md"
+requires-python = ">=3.10"
+license = {'file' = "LICENSE"}
+classifiers = [
+    "Programming Language :: Python :: 3",
+]
+dependencies = [
+    "torch >= 1.21.1",
+    "numpy >= 1.23.2",
+    "pandas >= 1.4.3",
+    "scipy >= 1.9.1",
+    "pyteomics >= 4.5.5",
+    "platformdirs"
+]
+
+[project.scripts]
+chronologer_train = "chronologer.train:main"
+chronologer_predict = "chronologer.predict:main"
+
+[tool.setuptools.packages.find]
+where = ["src"]
