+++ src/chronologer/chronologer_utils/custom.py
@@ -8,7 +8,7 @@ import os
 def return_custom_modifications():
     src_dir = os.path.dirname(os.path.abspath(__file__))
     custom_modifications = [ ]
-    for line in open( os.path.join( src_dir, '..', 'user_modifications.txt', ) ):
+    for line in open( os.path.join( src_dir, 'data', 'user_modifications.txt', ) ):
         if line[0] != '#':
             parts = line.rstrip().split('\t')
             assert len(parts) == 3, 'Invalid modification: '+line.rstrip()
