+++ src/chronologer/chronologer_utils/local_io.py
@@ -2,9 +2,9 @@ import os, sqlite3
 import numpy as np, pandas as pd
 from pyteomics.parser import cleave, expasy_rules
 
-import chronologer.src.constants as constants
-from chronologer.src.masses import masses
-from chronologer.src.tensorize import modseq_to_codedseq
+import chronologer.chronologer_utils.constants as constants
+from chronologer.chronologer_utils.masses import masses
+from chronologer.chronologer_utils.tensorize import modseq_to_codedseq
