+++ src/chronologer/training_functions.py
@@ -8,14 +8,14 @@ os.environ['KMP_DUPLICATE_LIB_OK']='True'
 import pandas as pd
 
 
-import chronologer.src.constants as constants 
-from chronologer.src.local_io import read_rt_database
-from chronologer.src.chronologer.settings import training_parameters
-from chronologer.src.tensorize import hi_db_to_tensors
-from chronologer.src.chronologer.model import initialize_chronologer_model
-from chronologer.src.loss_functions import RT_masked_negLogL, LogL_Loss
-from chronologer.src.training_loop import train_model
-from chronologer.src.utils import timer
+import chronologer.chronologer_utils.constants as constants 
+from chronologer.chronologer_utils.local_io import read_rt_database
+from chronologer.settings import training_parameters
+from chronologer.chronologer_utils.tensorize import hi_db_to_tensors
+from chronologer.model import initialize_chronologer_model
+from chronologer.chronologer_utils.loss_functions import RT_masked_negLogL, LogL_Loss
+from chronologer.chronologer_utils.training_loop import train_model
+from chronologer.chronologer_utils.utils import timer
 
 
 def rt_to_tensor_database( data_files, test_frac, random_seed, ):
